+++ packages/svelte/src/internal/client/dom/blocks/snippet.js
@@ -1,7 +1,6 @@
 /** @import { Snippet } from 'svelte' */
 /** @import { Effect, TemplateNode } from '#client' */
 /** @import { Getters } from '#shared' */
-import { add_snippet_symbol } from '../../../shared/validate.js';
 import { EFFECT_TRANSPARENT } from '../../constants.js';
 import { branch, block, destroy_effect, teardown } from '../../reactivity/effects.js';
 import {
@@ -11,6 +10,8 @@ import {
 import { hydrate_next, hydrate_node, hydrating } from '../hydration.js';
 import { create_fragment_from_html } from '../reconciler.js';
 import { assign_nodes } from '../template.js';
+import * as w from '../../warnings.js';
+import { DEV } from 'esm-env';
 
 /**
  * @template {(node: TemplateNode, ...args: any[]) => void} SnippetFn
@@ -53,7 +54,7 @@ export function snippet(node, get_snippet, ...args) {
  * @param {(node: TemplateNode, ...args: any[]) => void} fn
  */
 export function wrap_snippet(component, fn) {
-	return add_snippet_symbol((/** @type {TemplateNode} */ node, /** @type {any[]} */ ...args) => {
+	return (/** @type {TemplateNode} */ node, /** @type {any[]} */ ...args) => {
 		var previous_component_function = dev_current_component_function;
 		set_dev_current_component_function(component);
 
@@ -62,7 +63,7 @@ export function wrap_snippet(component, fn) {
 		} finally {
 			set_dev_current_component_function(previous_component_function);
 		}
-	});
+	};
 }
 
 /**
@@ -75,29 +76,33 @@ export function wrap_snippet(component, fn) {
  * @returns {Snippet<Params>}
  */
 export function createRawSnippet(fn) {
-	return add_snippet_symbol(
-		(/** @type {TemplateNode} */ anchor, /** @type {Getters<Params>} */ ...params) => {
-			var snippet = fn(...params);
+	// @ts-expect-error the types are a lie
+	return (/** @type {TemplateNode} */ anchor, /** @type {Getters<Params>} */ ...params) => {
+		var snippet = fn(...params);
 
-			/** @type {Element} */
-			var element;
+		/** @type {Element} */
+		var element;
 
-			if (hydrating) {
-				element = /** @type {Element} */ (hydrate_node);
-				hydrate_next();
-			} else {
-				var html = snippet.render().trim();
-				var fragment = create_fragment_from_html(html);
-				element = /** @type {Element} */ (fragment.firstChild);
-				anchor.before(element);
+		if (hydrating) {
+			element = /** @type {Element} */ (hydrate_node);
+			hydrate_next();
+		} else {
+			var html = snippet.render().trim();
+			var fragment = create_fragment_from_html(html);
+			element = /** @type {Element} */ (fragment.firstChild);
+
+			if (DEV && (element.nextSibling !== null || element.nodeType !== 3)) {
+				w.invalid_raw_snippet_render();
 			}
 
-			const result = snippet.setup?.(element);
-			assign_nodes(element, element);
-
-			if (typeof result === 'function') {
-				teardown(result);
-			}
+			anchor.before(element);
 		}
-	);
+
+		const result = snippet.setup?.(element);
+		assign_nodes(element, element);
+
+		if (typeof result === 'function') {
+			teardown(result);
+		}
+	};
 }