+++ packages/svelte/src/internal/client/index.js
@@ -1,4 +1,4 @@
-export { FILENAME, ORIGINAL } from '../../constants.js';
+export { FILENAME, HMR } from '../../constants.js';
 export { add_locations } from './dev/elements.js';
 export { hmr } from './dev/hmr.js';
 export {
@@ -164,9 +164,8 @@ export {
 export { snapshot } from '../shared/clone.js';
 export { noop } from '../shared/utils.js';
 export {
-	validate_component,
+	invalid_default_snippet,
 	validate_dynamic_element_tag,
-	validate_snippet,
 	validate_store,
 	validate_void_dynamic_element
 } from '../shared/validate.js';